Override setFragmentData instead of onCopy
